a2/requirements.txt
argparse==1.4.0
grpcio==1.25.0
grpcio-tools==1.25.0
sqlalchemy==1.3.10